[fluent] refactor FlyoutPositionProvider
