+++ fluent/src/commonMain/kotlin/com/konyaco/fluent/component/Flyout.kt
@@ -15,28 +15,16 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.Stable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.geometry.Size
-import androidx.compose.ui.geometry.center
 import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.input.key.KeyEvent
-import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.unit.Density
-import androidx.compose.ui.unit.IntOffset
-import androidx.compose.ui.unit.IntRect
-import androidx.compose.ui.unit.IntSize
-import androidx.compose.ui.unit.LayoutDirection
-import androidx.compose.ui.unit.center
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.PopupPositionProvider
 import androidx.compose.ui.window.PopupProperties
 import com.konyaco.fluent.animation.FluentDuration
 import com.konyaco.fluent.animation.FluentEasing
@@ -49,6 +37,7 @@ fun FlyoutContainer(
     modifier: Modifier = Modifier,
     initialVisible: Boolean = false,
     placement: FlyoutPlacement = FlyoutPlacement.Auto,
+    adaptivePlacement: Boolean = false,
     onKeyEvent: ((keyEvent: KeyEvent) -> Boolean)? = null,
     onPreviewKeyEvent: ((keyEvent: KeyEvent) -> Boolean)? = null,
     content: @Composable FlyoutScope.() -> Unit
@@ -59,6 +48,7 @@ fun FlyoutContainer(
                 visible = isFlyoutVisible,
                 onDismissRequest = { isFlyoutVisible = false },
                 placement = placement,
+                adaptivePlacement = adaptivePlacement,
                 onKeyEvent = onKeyEvent,
                 onPreviewKeyEvent = onPreviewKeyEvent,
                 content = flyout
@@ -112,6 +102,7 @@ fun Flyout(
     onDismissRequest: () -> Unit,
     modifier: Modifier = Modifier,
     placement: FlyoutPlacement = FlyoutPlacement.Auto,
+    adaptivePlacement: Boolean = false,
     shape: Shape = RoundedCornerShape(8.dp),
     onKeyEvent: ((keyEvent: KeyEvent) -> Boolean)? = null,
     onPreviewKeyEvent: ((keyEvent: KeyEvent) -> Boolean)? = null,
@@ -121,7 +112,10 @@ fun Flyout(
         visible = visible,
         onDismissRequest = onDismissRequest,
         modifier = modifier,
-        positionProvider = rememberFlyoutPositionProvider(initialPlacement = placement),
+        positionProvider = rememberFlyoutPositionProvider(
+            initialPlacement = placement,
+            adaptivePlacement = adaptivePlacement
+        ),
         shape = shape,
         onKeyEvent = onKeyEvent,
         onPreviewKeyEvent = onPreviewKeyEvent,
@@ -147,7 +141,7 @@ internal fun BasicFlyout(
     }
     visibleState.targetState = visible
     if (visibleState.currentState || visibleState.targetState) {
-        com.konyaco.fluent.component.Popup(
+        Popup(
             onDismissRequest = onDismissRequest,
             properties = PopupProperties(clippingEnabled = false, focusable = onKeyEvent != null || onPreviewKeyEvent != null),
             popupPositionProvider = positionProvider,
@@ -206,359 +200,6 @@ internal fun FlyoutContent(
                     content()
                 }
             }
-        }
-    }
-}
-
-@Composable
-internal fun rememberFlyoutPositionProvider(
-    initialPlacement: FlyoutPlacement = FlyoutPlacement.Auto,
-    paddingToAnchor: PaddingValues = PaddingValues(flyoutDefaultPadding)
-): FlyoutPositionProvider {
-    val density = LocalDensity.current
-    return remember(initialPlacement, density, paddingToAnchor) {
-        FlyoutPositionProvider(density, initialPlacement, paddingToAnchor)
-    }
-}
-
-@Stable
-internal open class FlyoutPositionProvider(
-    private val density: Density,
-    private val initialPlacement: FlyoutPlacement = FlyoutPlacement.Auto,
-    private val paddingToAnchor: PaddingValues = PaddingValues(flyoutDefaultPadding),
-) : PopupPositionProvider {
-
-    var applyAnimation by mutableStateOf(false)
-        private set
-
-    var targetPlacement by mutableStateOf(initialPlacement)
-        private set
-
-    override fun calculatePosition(
-        anchorBounds: IntRect,
-        windowSize: IntSize,
-        layoutDirection: LayoutDirection,
-        popupContentSize: IntSize
-    ): IntOffset {
-        applyAnimation = false
-        with(density) {
-            val popupPadding = flyoutPopPaddingFixShadowRender.toPx()
-            val popupActualSize = Size(
-                popupContentSize.width - popupPadding * 2,
-                popupContentSize.height - popupPadding * 2
-            )
-            val horizontalPlacement: HorizontalPlacement
-            val verticalPlacement: VerticalPlacement
-
-            targetPlacement = if (initialPlacement != FlyoutPlacement.Auto) {
-                initialPlacement.apply {
-                    val (targetHorizontalPlacement, targetVerticalPlacement) = transformPlacement(
-                        initialPlacement
-                    )
-                    horizontalPlacement = targetHorizontalPlacement
-                    verticalPlacement = targetVerticalPlacement
-                }
-            } else {
-                val (targetHorizontalPlacement, targetVerticalPlacement) = calculateTargetPlacement(
-                    anchorBounds,
-                    windowSize,
-                    layoutDirection,
-                    popupActualSize
-                )
-                horizontalPlacement = targetHorizontalPlacement
-                verticalPlacement = targetVerticalPlacement
-                transformPlacement(horizontalPlacement, verticalPlacement)
-            }
-            applyAnimation = true
-            if (targetPlacement == FlyoutPlacement.Full) {
-                return IntOffset((windowSize.width - popupContentSize.width) / 2, (windowSize.height - popupContentSize.height) / 2)
-            }
-            val popupActualCenter = popupActualSize.center
-            val anchorCenter = anchorBounds.center
-            return IntOffset(
-                x = (getOffsetX(
-                    horizontalPlacement,
-                    layoutDirection,
-                    anchorBounds,
-                    anchorCenter,
-                    popupActualSize,
-                    popupActualCenter,
-                    windowSize
-                ) - popupPadding).toInt(),
-                y = (getOffsetY(
-                    verticalPlacement,
-                    anchorBounds,
-                    anchorCenter,
-                    popupActualSize,
-                    popupActualCenter,
-                    windowSize
-                ) - popupPadding).toInt()
-            )
-        }
-    }
-
-    private fun Density.getOffsetX(
-        placement: HorizontalPlacement,
-        layoutDirection: LayoutDirection,
-        anchorBounds: IntRect,
-        anchorCenter: IntOffset,
-        popupContentSize: Size,
-        popupContentCenter: Offset,
-        windowSize: IntSize
-    ): Float {
-        val isLTR = layoutDirection == LayoutDirection.Ltr
-        return if (isLTR) {
-            when (placement) {
-                HorizontalPlacement.Start -> anchorBounds.left - popupContentSize.width - paddingToAnchor.calculateLeftPadding(
-                    layoutDirection
-                ).toPx()
-
-                HorizontalPlacement.End -> anchorBounds.right + paddingToAnchor.calculateRightPadding(
-                    layoutDirection
-                ).toPx()
-
-                HorizontalPlacement.Center -> anchorCenter.x - popupContentCenter.x
-                HorizontalPlacement.AlignedStart -> anchorBounds.left.toFloat()
-                HorizontalPlacement.AlignedEnd -> anchorBounds.right - popupContentSize.width
-                else -> windowSize.center.x - popupContentCenter.x
-            }
-        } else {
-            when (placement) {
-                HorizontalPlacement.End -> anchorBounds.left - popupContentSize.width - paddingToAnchor.calculateLeftPadding(
-                    layoutDirection
-                ).toPx()
-
-                HorizontalPlacement.Start -> anchorBounds.right + paddingToAnchor.calculateRightPadding(
-                    layoutDirection
-                ).toPx()
-
-                HorizontalPlacement.Center -> anchorCenter.x - popupContentCenter.x
-                HorizontalPlacement.AlignedEnd -> anchorBounds.left.toFloat()
-                HorizontalPlacement.AlignedStart -> anchorBounds.right - popupContentSize.width
-                else -> windowSize.center.x - popupContentCenter.x
-            }
-        }
-    }
-
-    private fun Density.getOffsetY(
-        placement: VerticalPlacement,
-        anchorBounds: IntRect,
-        anchorCenter: IntOffset,
-        popupContentSize: Size,
-        popupContentCenter: Offset,
-        windowSize: IntSize
-    ): Float {
-        return when (placement) {
-            VerticalPlacement.Top -> (anchorBounds.top - popupContentSize.height - paddingToAnchor.calculateTopPadding()
-                .toPx())
-
-            VerticalPlacement.Center -> anchorCenter.y - popupContentCenter.y
-            VerticalPlacement.Bottom -> anchorBounds.bottom + paddingToAnchor.calculateBottomPadding()
-                .toPx()
-
-            VerticalPlacement.AlignedTop -> anchorBounds.top.toFloat()
-            VerticalPlacement.AlignedBottom -> anchorBounds.bottom - popupContentSize.height
-            else -> windowSize.center.y - popupContentCenter.y
-        }
-    }
-
-    protected open fun Density.calculateTargetPlacement(
-        anchorBounds: IntRect,
-        windowSize: IntSize,
-        layoutDirection: LayoutDirection,
-        popupContentSize: Size
-    ): Pair<HorizontalPlacement, VerticalPlacement> {
-        return calculateVerticalPlacement(
-            anchorBounds,
-            windowSize,
-            layoutDirection,
-            popupContentSize
-        )
-    }
-
-    private fun Density.calculateVerticalPlacement(
-        anchorBounds: IntRect,
-        windowSize: IntSize,
-        layoutDirection: LayoutDirection,
-        popupContentSize: Size
-    ): Pair<HorizontalPlacement, VerticalPlacement> {
-        val hasTopSpace = anchorBounds.top - paddingToAnchor.calculateTopPadding()
-            .toPx() >= popupContentSize.height
-        val halfContentWidth = popupContentSize.width / 2f
-        val center = anchorBounds.center
-
-        fun calculateStartOrEndOrCenter(): HorizontalPlacement {
-            val hasLeftSpace = center.x >= halfContentWidth
-            val hasRightSpace = windowSize.width - center.x >= halfContentWidth
-            return when {
-                hasRightSpace && hasLeftSpace -> HorizontalPlacement.Center
-                (hasLeftSpace && layoutDirection == LayoutDirection.Ltr) or
-                        (hasRightSpace && layoutDirection == LayoutDirection.Rtl) -> HorizontalPlacement.Start
-
-                hasRightSpace -> HorizontalPlacement.Center
-                hasLeftSpace -> HorizontalPlacement.End
-                else -> HorizontalPlacement.None
-            }
-        }
-
-        if (hasTopSpace) {
-            return calculateStartOrEndOrCenter() to VerticalPlacement.Top
-        }
-        val hasBottomSpace = anchorBounds.bottom - paddingToAnchor.calculateBottomPadding()
-            .toPx() >= popupContentSize.height
-
-        if (hasBottomSpace) {
-            return calculateStartOrEndOrCenter() to VerticalPlacement.Bottom
-        }
-        return HorizontalPlacement.Center to VerticalPlacement.Top
-    }
-
-    protected fun Density.calculateHorizontalPlacement(
-        anchorBounds: IntRect,
-        windowSize: IntSize,
-        layoutDirection: LayoutDirection,
-        popupContentSize: Size
-    ): Pair<HorizontalPlacement, VerticalPlacement> {
-        val isLRT = layoutDirection == LayoutDirection.Ltr
-        val halfContentWidth = popupContentSize.width / 2f
-        val halfContentHeight = popupContentSize.height / 2f
-        val center = anchorBounds.center
-        val hasLeftSpace =
-            center.x >= halfContentWidth + paddingToAnchor.calculateLeftPadding(layoutDirection)
-                .toPx()
-        val hasRightSpace =
-            windowSize.width - center.x >= halfContentWidth + paddingToAnchor.calculateRightPadding(
-                layoutDirection
-            ).toPx()
-
-        fun calculateTopOrBottomOrCenter(): VerticalPlacement {
-            val hasAlignedTopSpace = windowSize.height - anchorBounds.top >= popupContentSize.height
-            val hasAlignedBottomSpace = anchorBounds.bottom >= popupContentSize.height
-            if (hasAlignedTopSpace) {
-                return VerticalPlacement.AlignedTop
-            }
-            if (hasAlignedBottomSpace) {
-                return VerticalPlacement.AlignedBottom
-            }
-            if (center.y - halfContentHeight > 0) {
-                return VerticalPlacement.Center
-            }
-            return VerticalPlacement.None
-        }
-
-        if (isLRT) {
-            if (hasRightSpace) {
-                return HorizontalPlacement.End to calculateTopOrBottomOrCenter()
-            }
-
-            if (hasLeftSpace) {
-                return HorizontalPlacement.Start to calculateTopOrBottomOrCenter()
-            }
-        } else {
-            if (hasLeftSpace) {
-                return HorizontalPlacement.End to calculateTopOrBottomOrCenter()
-            }
-            if (hasRightSpace) {
-                return HorizontalPlacement.Start to calculateTopOrBottomOrCenter()
-            }
-        }
-
-        return HorizontalPlacement.End to VerticalPlacement.AlignedTop
-    }
-
-    private fun transformPlacement(
-        horizontalPlacement: HorizontalPlacement,
-        verticalPlacement: VerticalPlacement
-    ): FlyoutPlacement {
-        return when (horizontalPlacement) {
-            HorizontalPlacement.Start -> {
-                when (verticalPlacement) {
-                    VerticalPlacement.AlignedTop -> FlyoutPlacement.StartAlignedTop
-                    VerticalPlacement.AlignedBottom -> FlyoutPlacement.StartAlignedBottom
-                    else -> FlyoutPlacement.Start
-                }
-            }
-
-            HorizontalPlacement.End -> {
-                when (verticalPlacement) {
-                    VerticalPlacement.AlignedTop -> FlyoutPlacement.EndAlignedTop
-                    VerticalPlacement.AlignedBottom -> FlyoutPlacement.EndAlignedBottom
-                    else -> FlyoutPlacement.End
-                }
-            }
-
-            HorizontalPlacement.Center -> {
-                when (verticalPlacement) {
-                    VerticalPlacement.Top -> FlyoutPlacement.Top
-                    VerticalPlacement.Bottom -> FlyoutPlacement.Bottom
-                    else -> FlyoutPlacement.Full
-                }
-            }
-
-            HorizontalPlacement.AlignedStart -> {
-                when (verticalPlacement) {
-                    VerticalPlacement.Top -> FlyoutPlacement.TopAlignedStart
-                    VerticalPlacement.Bottom -> FlyoutPlacement.BottomAlignedStart
-                    else -> FlyoutPlacement.Full
-                }
-            }
-
-            HorizontalPlacement.AlignedEnd -> {
-                when (verticalPlacement) {
-                    VerticalPlacement.Top -> FlyoutPlacement.TopAlignedEnd
-                    VerticalPlacement.Bottom -> FlyoutPlacement.BottomAlignedEnd
-                    else -> FlyoutPlacement.Full
-                }
-            }
-
-            else -> FlyoutPlacement.Full
-        }
-    }
-
-    private fun transformPlacement(flyoutPlacement: FlyoutPlacement): Pair<HorizontalPlacement, VerticalPlacement> {
-        return when (flyoutPlacement) {
-            FlyoutPlacement.Top -> HorizontalPlacement.Center to VerticalPlacement.Top
-            FlyoutPlacement.Bottom -> HorizontalPlacement.Center to VerticalPlacement.Bottom
-            FlyoutPlacement.Start -> HorizontalPlacement.Start to VerticalPlacement.Center
-            FlyoutPlacement.End -> HorizontalPlacement.End to VerticalPlacement.Center
-            FlyoutPlacement.TopAlignedStart -> HorizontalPlacement.AlignedStart to VerticalPlacement.Top
-            FlyoutPlacement.TopAlignedEnd -> HorizontalPlacement.AlignedEnd to VerticalPlacement.Top
-            FlyoutPlacement.BottomAlignedStart -> HorizontalPlacement.AlignedStart to VerticalPlacement.Bottom
-            FlyoutPlacement.BottomAlignedEnd -> HorizontalPlacement.AlignedEnd to VerticalPlacement.Bottom
-            FlyoutPlacement.StartAlignedTop -> HorizontalPlacement.Start to VerticalPlacement.AlignedTop
-            FlyoutPlacement.StartAlignedBottom -> HorizontalPlacement.Start to VerticalPlacement.AlignedBottom
-            FlyoutPlacement.EndAlignedTop -> HorizontalPlacement.End to VerticalPlacement.AlignedTop
-            FlyoutPlacement.EndAlignedBottom -> HorizontalPlacement.End to VerticalPlacement.AlignedBottom
-            FlyoutPlacement.Full -> HorizontalPlacement.Center to VerticalPlacement.Center
-            else -> HorizontalPlacement.None to VerticalPlacement.None
-        }
-    }
-
-    @JvmInline
-    protected value class HorizontalPlacement(
-        private val value: Int
-    ) {
-        companion object {
-            val Start = HorizontalPlacement(0)
-            val Center = HorizontalPlacement(1)
-            val End = HorizontalPlacement(2)
-            val None = HorizontalPlacement(-1)
-            val AlignedStart = HorizontalPlacement(3)
-            val AlignedEnd = HorizontalPlacement(4)
-        }
-    }
-
-    @JvmInline
-    protected value class VerticalPlacement(
-        private val value: Int
-    ) {
-        companion object {
-            val Top = VerticalPlacement(0)
-            val Center = VerticalPlacement(1)
-            val Bottom = VerticalPlacement(2)
-            val AlignedTop = VerticalPlacement(3)
-            val AlignedBottom = VerticalPlacement(4)
-            val None = VerticalPlacement(-1)
         }
     }
 }
